pyproject.toml
[tool.poetry]
name = "yandex-alice-quest-dictionary"
version = "0.1.0"
description = ""
authors = ["StanisLove <sta-tun-slav@yandex.ru>"]
readme = "README.md"
packages = [{include = "yandex_alice_quest_dictionary"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
python-dotenv = "^0.21.1"
pymongo = "^4.3.3"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"
typer = "^0.7.0"

[tool.poetry.scripts]
start = "yandex_alice_quest_dictionary.main:run"
setup_db = "yandex_alice_quest_dictionary.db.mongo:setup_dicitionaries"
get_data = "yandex_alice_quest_dictionary.db.mongo:get_data"
drop_db = "yandex_alice_quest_dictionary.db.mongo:drop_dictionaries"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
